pyproject.toml
[project]
name = "serial-hid-kvm"
version = "0.1.0"
description = "KVM control via CH9329 USB HID emulator + HDMI capture with interactive preview and TCP API"
requires-python = ">=3.10"
dependencies = [
    "pyserial>=3.5",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "pynput>=1.7.6",
    "pyyaml>=6.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
audio = ["sounddevice>=0.4.6"]

[project.scripts]
serial-hid-kvm = "serial_hid_kvm.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
